Add setting for toggling included package styles
Note that the implementation uses undocumented Atom API calls, which
might not be resilient to future updates. However, the approach taken
will let users see the effect of having package styles added or
removed without having to reload Atom, which is often the advice given
about packages that deal with styling.