Add match rules to style alternative documentation blocks
The difficulty with styling /** and /*! documentation blocks is that
they span over multiple lines with no definite leading tokens.
Inside these kinds of blocks, the added match rules will only recognise
Markdown constructs on lines with leading asterisks.

It appears rustdoc also has a hard time recognising certain constructs,
like unordered list items, when a leading asterisk is not present in
this type of a documentation block.

Also fixes up /// and //! documentation blocks to be recognised in places
other than the start of a new line.