Revise primitive type match rule
Any module name that happens to share the same name as a primitive
type will no longer be matched. The match rule now takes into account
whether a potential primitive type match is being used as a path.

For example, `str` in `std::str::from_utf8` will not be matched.

In addition, this change removes match rules for identifying raw
pointers as well as the `String` struct.