Remove lambda match rule
It was initially thought that all forms of lambda expression usage were
manageable with match rules. After much observation of Rust code in the
wild, this thought appears to be an oversight. There is only so much that
can be done with regular expressions until it becomes a futile effort.

Distinguishing lambda expressions from uses of logical OR, or bitwise
operator expressions, or even from within a function as an argument,
is quite plainly difficult because of bracketing, among other formatting
cases.