Fix toggle line comments editor command
The Atom editor contains a handy shortcut for toggling text as code
comments, which depends on the `commentStart` and `commentEnd` properties
defined in the settings of a package. The comment toggling appears to only
work for entire lines.

Issue #3 points out that the toggle shortcut does not currently behave in
a correct fashion. The fix simply involves removing the superfluous
`commentEnd` property and changes the `commentStart` to be a plain line
comment delimiter in Rust.

Also adds some tests to cover this comment toggling feature.