Rework macros-related match rules
Replaces the "macros-common" match group with a simpler one named
"metavariables". This new match group is included as a common match group
so that usage of variables specific to macros will be matched in areas
like item definitions.

These changes also ensure the dollar sign notation for macros is tokenised
as a symbol.