Expose configuration for token colours
Makes available an additional package setting for customising the colour
of symbols (i.e. curly braces, commas) and operators (i.e. equals, plus).
In doing so, the implementation is a bit of a hack as it relies on an
unused grammar scope named 'token' for linking the package setting colour
with that of the style sheet used.