Fix matching for where clauses across multiple lines
This fix addresses issue #6 to ensure that the `where` keyword is
recognised when its part of a clause that begins on another line.

The previous opening match rule logic required exactly one whitespace
character to precede the keyword, which will always fail if the
`where` keyword commences a line. This matching requirement is now relaxed
in favour of word boundaries on either side of the `where` keyword.