Update call for reading default syntax colour values
The call to compile the package stylesheet for default syntax colour
values appears to throw errors when supplied `null` as a first argument on
some distributions of Atom (e.g. via community repository for ArchLinux).

It has now been modified to accept the actual path of the package
stylesheet. In retrospect, this is probably the more correct thing to do.

This change addresses #8.