Add match rule for untagged union declarations
Highlighting of untagged union declarations has been restricted to a
single line. The reason is due to expectations of backward compatibility
where `union` can be used as a binding name.

Making declarations work multi-line is not impossible, it would just
require covering all usages of `union` where it is not a keyword.

Refer to RFC 1444 for details on the feature.