Create new match rule for the let keyword
The new match rule is intended to isolate the name or pattern involved in
making variable bindings.

All existing match rules which simply matched the `let` keyword by name
have been updated to no longer do so because they are no longer necessary.

This change should also help prevent false positives, like matching uses
of `union` as a name and not as the keyword.