Add dedent-js dependency to help test fixtures
Existing usages of template string literals as test case fixtures do not
appear to be sensitive to indentation. For test cases where indentation
matters, the `dedent-js` package should be of use in stripping out leading
whitespace for token matching.

Dependency was chosen fairly arbitrarily. It is lean and does the job.